backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
chromadb==0.5.5
pypdf==4.3.1
python-dotenv==1.0.1
openai==1.45.0
pydantic==2.8.2
